generalize function mapping to collections
